+++ flink/v1.19/flink/src/main/java/org/apache/iceberg/flink/sink/IcebergCommittable.java
@@ -16,12 +16,11 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.iceberg.flink.sink.committer;
+package org.apache.iceberg.flink.sink;
 
 import java.io.Serializable;
 import java.util.Arrays;
 import java.util.Objects;
-import org.apache.flink.annotation.Internal;
 import org.apache.iceberg.relocated.com.google.common.base.MoreObjects;
 
 /**
@@ -32,33 +31,32 @@ import org.apache.iceberg.relocated.com.google.common.base.MoreObjects;
  * <p>{@link IcebergCommittableSerializer} is used for serializing the objects between the Writer
  * and the Aggregator operator and between the Aggregator and the Committer as well.
  */
-@Internal
-public class IcebergCommittable implements Serializable {
+class IcebergCommittable implements Serializable {
   private final byte[] manifest;
   private final String jobId;
   private final String operatorId;
   private final long checkpointId;
 
-  public IcebergCommittable(byte[] manifest, String jobId, String operatorId, long checkpointId) {
+  IcebergCommittable(byte[] manifest, String jobId, String operatorId, long checkpointId) {
     this.manifest = manifest;
     this.jobId = jobId;
     this.operatorId = operatorId;
     this.checkpointId = checkpointId;
   }
 
-  public byte[] manifest() {
+  byte[] manifest() {
     return manifest;
   }
 
-  public String jobId() {
+  String jobId() {
     return jobId;
   }
 
-  public String operatorId() {
+  String operatorId() {
     return operatorId;
   }
 
-  public Long checkpointId() {
+  Long checkpointId() {
     return checkpointId;
   }
