+++ flink/v1.19/flink/src/test/java/org/apache/iceberg/flink/sink/TestIcebergCommitter.java
@@ -84,10 +84,6 @@ import org.apache.iceberg.flink.FlinkSchemaUtil;
 import org.apache.iceberg.flink.SimpleDataUtil;
 import org.apache.iceberg.flink.TableLoader;
 import org.apache.iceberg.flink.TestHelpers;
-import org.apache.iceberg.flink.sink.committer.IcebergCommittable;
-import org.apache.iceberg.flink.sink.committer.IcebergCommittableSerializer;
-import org.apache.iceberg.flink.sink.committer.IcebergCommitter;
-import org.apache.iceberg.flink.sink.committer.IcebergWriteAggregator;
 import org.apache.iceberg.io.FileAppenderFactory;
 import org.apache.iceberg.io.WriteResult;
 import org.apache.iceberg.relocated.com.google.common.collect.ImmutableList;
